requirements.txt
customtkinter==5.2.2
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
schedule==1.2.1
Pillow==10.1.0
pyinstaller==6.3.0
